requirements.txt
dnspython
nmap
python-nmap
requests